Refactor socsec module to allow import it from external projects.
socsec was designed in a way it could be used as a standalone app.
However, it should be possible to import it also as a module.
Keep only __version__ in socsec/__init__.py, move the rest to the
socsec/tools.py file. Fix imports, allow setup.py to understand and
signal dependencies of this module.

With these changes we should be able to install socsec as a dependency
for other projects using e.g. poetry.